refactor: move IconInfo type + methods out of icons_map.go data file
